+++ packages/atproto/lib/src/entities/subscribed_repo_commit.g.dart
@@ -22,6 +22,8 @@ _$_SubscribedRepoCommit _$$_SubscribedRepoCommitFromJson(Map json) =>
                   .toList()),
           did: $checkedConvert('repo', (v) => v as String),
           cursor: $checkedConvert('seq', (v) => v as int),
+          rev: $checkedConvert('rev', (v) => v as String),
+          since: $checkedConvert('since', (v) => v as String?),
           isRebase: $checkedConvert('rebase', (v) => v as bool),
           isTooBig: $checkedConvert('tooBig', (v) => v as bool),
           createdAt:
@@ -44,6 +46,8 @@ Map<String, dynamic> _$$_SubscribedRepoCommitToJson(
       'ops': instance.ops.map((e) => e.toJson()).toList(),
       'repo': instance.did,
       'seq': instance.cursor,
+      'rev': instance.rev,
+      'since': instance.since,
       'rebase': instance.isRebase,
       'tooBig': instance.isTooBig,
       'time': instance.createdAt.toIso8601String(),